chore(ecs-mcp-server): renaming test files
